refactor: ticket interaction route table (v3.1.10 part 1/7)
Replace the 700+ LOC `handleTicketInteraction` if-ladder with a route
table. `src/events/ticketInteraction.ts` shrinks from 708 → 7 LOC and
delegates to a new dispatcher in `src/events/ticket/interactionRoutes.ts`
that keys on `customId` (exact and prefix) per interaction type
(button / modal / select).

Each branch body is now a dedicated handler in its own file:
- `create.ts` — createTicketButton, cancelTicketButton, selectTicketType,
  legacyTicketTypeButton, submitTicketModal (+ legacy modal/description
  helpers moved from ticketInteraction.ts)
- `typeAdmin.ts` — pingToggleButton (for `ticket_type_ping_toggle:*`)
- `adminOnly.ts` — adds adminOnlyButton, confirmAdminOnly, cancelAdminOnly
  (next to existing ticketAdminOnlyEvent)
- `close.ts` — adds closeButton, confirmClose, cancelClose
  (next to existing ticketCloseEvent)

The 3 already-extracted modal handlers (typeAdd, typeEdit, emailImport)
are routed via tiny adapters in interactionRoutes.ts so all route values
share the `(client, interaction)` signature.

Drops the dead `if (!guildId)` branch at the old ticket_type_ping_toggle
site that was already guarded by the early return at the top of the
function (logic_clarity finding).

Behavior unchanged. 1071 tests still passing. Build + biome clean.
