+++ src/events/ticketInteraction.ts
@@ -1,708 +1,7 @@
-import {
-  ActionRowBuilder,
-  ButtonBuilder,
-  ButtonStyle,
-  type Client,
-  type GuildMember,
-  type Interaction,
-  MessageFlags,
-  ModalBuilder,
-  type ModalSubmitFields,
-  type TextChannel,
-  TextInputBuilder,
-  TextInputStyle,
-} from 'discord.js';
-import { emailImportModalHandler } from '../commands/handlers/ticket/emailImport';
-import { buildTypeConfirmationEmbed, typeAddModalHandler } from '../commands/handlers/ticket/typeAdd';
-import { typeEditModalHandler } from '../commands/handlers/ticket/typeEdit';
-import { BotConfig } from '../typeorm/entities/BotConfig';
-import { StaffRole } from '../typeorm/entities/StaffRole';
-import { CustomTicketType } from '../typeorm/entities/ticket/CustomTicketType';
-import { Ticket } from '../typeorm/entities/ticket/Ticket';
-import { TicketConfig } from '../typeorm/entities/ticket/TicketConfig';
-import { UserTicketRestriction } from '../typeorm/entities/ticket/UserTicketRestriction';
-import {
-  createPrivateChannelPermissions,
-  createRateLimitKey,
-  enhancedLogger,
-  escapeDiscordMarkdown,
-  extractIdFromMention,
-  LogCategory,
-  lang,
-  PermissionSets,
-  RateLimits,
-  rateLimiter,
-} from '../utils';
-import { lazyRepo } from '../utils/database/lazyRepo';
-import { isLegacyTicketType, resolveLegacyPingColumn, resolveTicketType } from '../utils/ticket/legacyTypes';
-import { customTicketOptions, ticketOptions } from './ticket';
-import { ticketAdminOnlyEvent } from './ticket/adminOnly';
-import { ageVerifyMessage, ageVerifyModal } from './ticket/ageVerify';
-import { banAppealMessage, banAppealModal } from './ticket/banAppeal';
-import { bugReportMessage, bugReportModal } from './ticket/bugReport';
-import { ticketCloseEvent } from './ticket/close';
-import { otherMessage, otherModal } from './ticket/other';
-import { playerReportMessage, playerReportModal } from './ticket/playerReport';
-
-/** Build a legacy ticket modal with the correct inputs for the given type. */
-function buildLegacyTicketModal(typeId: string, modal: ModalBuilder): ModalBuilder {
-  switch (typeId) {
-    case '18_verify':
-      return ageVerifyModal(modal);
-    case 'ban_appeal':
-      return banAppealModal(modal);
-    case 'player_report':
-      return playerReportModal(modal);
-    case 'bug_report':
-      return bugReportModal(modal);
-    case 'other':
-      return otherModal(modal);
-    default:
-      return modal;
-  }
-}
-
-/** Build legacy ticket description from modal submit fields. */
-function buildLegacyTicketDescription(typeId: string, fields: ModalSubmitFields): string {
-  switch (typeId) {
-    case '18_verify':
-      return ageVerifyMessage(fields);
-    case 'ban_appeal':
-      return banAppealMessage(fields);
-    case 'player_report':
-      return playerReportMessage(fields);
-    case 'bug_report':
-      return bugReportMessage(fields);
-    case 'other':
-      return otherMessage(fields);
-    default:
-      return '';
-  }
-}
-
-const ticketConfigRepo = lazyRepo(TicketConfig);
-const ticketRepo = lazyRepo(Ticket);
-const staffRoleRepo = lazyRepo(StaffRole);
-const botConfigRepo = lazyRepo(BotConfig);
-const customTypeRepo = lazyRepo(CustomTicketType);
+import type { Client, Interaction } from 'discord.js';
+import { dispatchTicketInteraction } from './ticket/interactionRoutes';
 
 export const handleTicketInteraction = async (client: Client, interaction: Interaction) => {
   if (!interaction.guildId) return;
-  const guildId = interaction.guildId;
-  // ticketConfig is loaded lazily — only fetched when actually needed
-  let ticketConfig: import('../typeorm/entities/ticket/TicketConfig').TicketConfig | null = null;
-  const getTicketConfig = async () => {
-    if (ticketConfig === null) {
-      ticketConfig = await ticketConfigRepo.findOneBy({ guildId });
-    }
-    return ticketConfig;
-  };
-
-  /* Handle Custom Ticket Type Modals */
-  if (interaction.isModalSubmit()) {
-    if (interaction.customId === 'ticket-type-add-modal') {
-      await typeAddModalHandler(interaction);
-      return;
-    }
-
-    if (interaction.customId.startsWith('ticket-type-edit-modal:')) {
-      const typeId = interaction.customId.replace('ticket-type-edit-modal:', '');
-      await typeEditModalHandler(interaction, typeId);
-      return;
-    }
-
-    if (interaction.customId === 'ticket-email-import-modal') {
-      await emailImportModalHandler(interaction);
-      return;
-    }
-  }
-
-  /* Handle Ticket Type Ping Toggle Button */
-  if (interaction.isButton() && interaction.customId.startsWith('ticket_type_ping_toggle:')) {
-    const typeId = interaction.customId.replace('ticket_type_ping_toggle:', '');
-    enhancedLogger.debug(`Button: staff ping toggle for type '${typeId}'`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-      typeId,
-    });
-
-    if (!guildId) {
-      enhancedLogger.warn('Staff ping toggle failed: guild not found', LogCategory.COMMAND_EXECUTION, {
-        userId: interaction.user.id,
-      });
-      await interaction.reply({
-        content: lang.general.cmdGuildNotFound,
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-
-    const type = await customTypeRepo.findOne({
-      where: { guildId, typeId },
-    });
-
-    if (!type) {
-      enhancedLogger.warn(`Staff ping toggle failed: type '${typeId}' not found`, LogCategory.COMMAND_EXECUTION, {
-        userId: interaction.user.id,
-        guildId,
-        typeId,
-      });
-      await interaction.reply({
-        content: lang.ticket.customTypes.typeEdit.notFound,
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-
-    // Toggle the ping setting
-    const previousState = type.pingStaffOnCreate;
-    type.pingStaffOnCreate = !type.pingStaffOnCreate;
-    await customTypeRepo.save(type);
-    enhancedLogger.info(
-      `Staff ping toggled for type '${typeId}': ${previousState} → ${type.pingStaffOnCreate}`,
-      LogCategory.COMMAND_EXECUTION,
-      {
-        userId: interaction.user.id,
-        guildId,
-        typeId,
-        previousState,
-        newState: type.pingStaffOnCreate,
-      },
-    );
-
-    // Rebuild the embed with updated state
-    const embed = buildTypeConfirmationEmbed(type, false);
-
-    // Update the button to reflect the new state
-    const tl = lang.ticket.customTypes.typeAdd;
-    const toggleButton = new ButtonBuilder()
-      .setCustomId(`ticket_type_ping_toggle:${typeId}`)
-      .setLabel(type.pingStaffOnCreate ? tl.pingToggleDisable : tl.pingToggleEnable)
-      .setStyle(type.pingStaffOnCreate ? ButtonStyle.Danger : ButtonStyle.Success)
-      .setEmoji(type.pingStaffOnCreate ? '🔕' : '🔔');
-
-    const buttonRow = new ActionRowBuilder<ButtonBuilder>().addComponents(toggleButton);
-
-    // Update the message with the new embed and button state
-    await interaction.update({
-      embeds: [embed],
-      components: [buttonRow],
-    });
-
-    return;
-  }
-
-  /* Create Ticket Button */
-  if (interaction.isButton() && interaction.customId === 'create_ticket') {
-    enhancedLogger.debug(`Button: create_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-
-    // check if the ticket config exists
-    const config = await getTicketConfig();
-    if (!config) {
-      enhancedLogger.warn('Create ticket failed: ticketConfig not found', LogCategory.COMMAND_EXECUTION, {
-        userId: interaction.user.id,
-        guildId,
-      });
-      return;
-    }
-
-    // check if we have the right messageid
-    if (config.messageId === interaction.message.id) {
-      try {
-        // Try to get custom ticket types (filtered by user restrictions)
-        const customOptions = await customTicketOptions(guildId, interaction.user.id);
-        await interaction.reply({
-          content: lang.ticket.selectTicketType,
-          components: [customOptions],
-          flags: [MessageFlags.Ephemeral],
-        });
-      } catch (error) {
-        // Fallback to legacy buttons if custom types fail
-        enhancedLogger.warn('Failed to load custom ticket types, using legacy options', LogCategory.COMMAND_EXECUTION, {
-          guildId,
-          error: error instanceof Error ? error.message : String(error),
-        });
-        const options = ticketOptions();
-        await interaction.reply({
-          content: lang.ticket.selectTicketType,
-          components: [options],
-          flags: [MessageFlags.Ephemeral],
-        });
-      }
-    }
-  }
-
-  /* Cancel Ticket Button */
-  if (interaction.isButton() && interaction.customId === 'cancel_ticket') {
-    enhancedLogger.debug(`Button: cancel_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-
-    // Update the message to remove components and show cancellation
-    await interaction.update({
-      content: lang.ticket.cancelled,
-      components: [],
-    });
-  }
-
-  /* Custom Ticket Type Select Menu */
-  if (interaction.isStringSelectMenu() && interaction.customId === 'ticket_type_select') {
-    const selectedTypeId = interaction.values[0];
-    enhancedLogger.debug(`Select: ticket type '${selectedTypeId}'`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-      selectedTypeId,
-    });
-
-    // Handle "none" option (user has no available ticket types)
-    if (selectedTypeId === 'none') {
-      await interaction.reply({
-        content: '🚫 You do not have access to create any ticket types.',
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-
-    // Check if user is restricted from creating this ticket type
-    const restrictionRepo = lazyRepo(UserTicketRestriction);
-    const restriction = await restrictionRepo.findOne({
-      where: { guildId, userId: interaction.user.id, typeId: selectedTypeId },
-    });
-
-    if (restriction) {
-      await interaction.reply({
-        content: '🚫 You are not allowed to create this type of ticket.',
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-
-    // Check if this is a LEGACY ticket type that should use hardcoded modals
-    if (isLegacyTicketType(selectedTypeId)) {
-      enhancedLogger.debug(`Opening legacy modal for type: ${selectedTypeId}`, LogCategory.COMMAND_EXECUTION, {
-        userId: interaction.user.id,
-        guildId,
-        ticketType: selectedTypeId,
-      });
-
-      // Use legacy modal builders for legacy types
-      const modal = buildLegacyTicketModal(
-        selectedTypeId,
-        new ModalBuilder()
-          .setCustomId(`ticket_modal_${selectedTypeId}`)
-          .setTitle(`Create ${selectedTypeId.replace('_', ' ')} Ticket`),
-      );
-
-      // Show the modal (this is the ONLY response we can give)
-      await interaction.showModal(modal);
-
-      return;
-    }
-
-    // Get the custom ticket type details
-    const typeRepo = customTypeRepo;
-    const ticketType = await typeRepo.findOne({
-      where: { guildId, typeId: selectedTypeId },
-    });
-
-    if (!ticketType) {
-      await interaction.reply({
-        content: '❌ Selected ticket type not found!',
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-
-    // Build modal with custom fields or default description
-    const modal = new ModalBuilder()
-      .setCustomId(`ticket_modal_${ticketType.typeId}`)
-      .setTitle(`${ticketType.emoji || '🎫'} ${ticketType.displayName}`);
-
-    // Check if custom fields are configured
-    if (ticketType.customFields && ticketType.customFields.length > 0) {
-      // Use custom fields (max 5 fields per modal in Discord)
-      const fieldsToAdd = ticketType.customFields.slice(0, 5);
-
-      for (const field of fieldsToAdd) {
-        const input = new TextInputBuilder()
-          .setCustomId(field.id)
-          .setLabel(field.label)
-          .setStyle(field.style === 'short' ? TextInputStyle.Short : TextInputStyle.Paragraph)
-          .setRequired(field.required);
-
-        if (field.placeholder) input.setPlaceholder(field.placeholder);
-        if (field.minLength) input.setMinLength(field.minLength);
-        if (field.maxLength) input.setMaxLength(field.maxLength);
-
-        const actionRow = new ActionRowBuilder<TextInputBuilder>().addComponents(input);
-        modal.addComponents(actionRow);
-      }
-    } else {
-      // No custom fields - use default description field
-      const descriptionInput = new TextInputBuilder()
-        .setCustomId('ticket_description')
-        .setLabel('Please describe your issue')
-        .setStyle(TextInputStyle.Paragraph)
-        .setPlaceholder(ticketType.description || 'Provide details about your ticket...')
-        .setRequired(true)
-        .setMaxLength(2000);
-
-      const actionRow = new ActionRowBuilder<TextInputBuilder>().addComponents(descriptionInput);
-      modal.addComponents(actionRow);
-    }
-
-    // Show the modal (this consumes the interaction)
-    await interaction.showModal(modal);
-
-    // Delete the ephemeral message after a short delay
-    // We need to wait a bit for the modal to fully open
-    setTimeout(async () => {
-      try {
-        // Delete the original ephemeral message
-        await interaction.message.delete();
-      } catch {
-        // Silently fail - message might already be gone
-      }
-    }, 500);
-
-    return;
-  }
-
-  /* Ticket Option Buttons */
-  if (interaction.isButton() && interaction.customId.startsWith('ticket_')) {
-    const ticketType = interaction.customId.replace('ticket_', '');
-
-    // Only handle valid ticket types (ignore bot setup buttons like ticket_skip, ticket_enable)
-    if (!isLegacyTicketType(ticketType)) {
-      return; // Not a ticket creation button, ignore it
-    }
-
-    enhancedLogger.debug(`Button: ticket_${ticketType}`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-      ticketType,
-    });
-
-    // build a modal for user input using shared legacy builder
-    const modal = buildLegacyTicketModal(
-      ticketType,
-      new ModalBuilder()
-        .setCustomId(`ticket_modal_${ticketType}`)
-        .setTitle(`Create ${ticketType.replace('_', ' ')} Ticket`),
-    );
-
-    await interaction.showModal(modal);
-  }
-
-  // handle ticket modal submission
-  if (interaction.isModalSubmit() && interaction.customId.startsWith('ticket_modal_')) {
-    const ticketType = interaction.customId.replace('ticket_modal_', '');
-    const member = interaction.member as GuildMember;
-    const guild = interaction.guild;
-    const modalTicketConfig = await getTicketConfig();
-    const category = modalTicketConfig?.categoryId;
-
-    enhancedLogger.debug(`Modal submit: ticket_modal_${ticketType}`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-      ticketType,
-    });
-
-    if (!guild) {
-      await interaction.reply({
-        content: lang.general.cmdGuildNotFound,
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-
-    if (!category) {
-      await interaction.reply({
-        content: lang.ticket.ticketCategoryNotFound,
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-
-    // Check rate limit (3 tickets per hour per user)
-    const rateLimitKey = createRateLimitKey.user(interaction.user.id, 'ticket-create');
-    const rateCheck = rateLimiter.check(rateLimitKey, RateLimits.TICKET_CREATE);
-
-    if (!rateCheck.allowed) {
-      await interaction.reply({
-        content: rateCheck.message,
-        flags: [MessageFlags.Ephemeral],
-      });
-      enhancedLogger.warn(`User hit ticket creation rate limit`, LogCategory.SECURITY, {
-        userId: interaction.user.id,
-        guildId,
-      });
-      return;
-    }
-
-    try {
-      // get user input from modal
-      const fields = interaction.fields;
-      let description = '';
-
-      // Resolve ticket type once — unified shape for legacy and custom types
-      const resolved = await resolveTicketType(guildId, ticketType);
-
-      if (!resolved) {
-        await interaction.reply({
-          content: '❌ Ticket type configuration not found!',
-          flags: [MessageFlags.Ephemeral],
-        });
-        return;
-      }
-
-      const isLegacyType = resolved.isLegacy;
-      const displayName = resolved.displayName || ticketType;
-
-      if (isLegacyType) {
-        description = buildLegacyTicketDescription(ticketType, fields);
-      } else {
-        // customType is guaranteed non-null when isLegacy is false
-        const customTypeConfig = resolved.customType;
-        const header = `# ${displayName}\n`;
-
-        if (customTypeConfig?.customFields && customTypeConfig.customFields.length > 0) {
-          // Build formatted description from all custom field responses
-          const fieldResponses: string[] = [];
-
-          for (const field of customTypeConfig.customFields) {
-            try {
-              const value = fields.getTextInputValue(field.id);
-              fieldResponses.push(`**${field.label}:** ${escapeDiscordMarkdown(value)}`);
-            } catch {
-              // Field may not be present in the modal submission — skip silently
-            }
-          }
-
-          description = header + fieldResponses.join('\n');
-        } else {
-          // No custom fields - use default description field
-          const defaultValue = fields.getTextInputValue('ticket_description');
-          description = header + defaultValue;
-        }
-      }
-
-      // create new ticket in the database
-      const ticketData: Partial<Ticket> = {
-        guildId: guildId,
-        createdBy: interaction.user.id,
-        type: ticketType,
-      };
-
-      if (!isLegacyType) {
-        ticketData.customTypeId = ticketType;
-      }
-
-      const newTicket = ticketRepo.create(ticketData);
-      const savedTicket = (await ticketRepo.save(newTicket)) as Ticket;
-
-      // create the ticket channel with numbering (sanitized for Discord channel names)
-      const sanitizedDisplayName = displayName.toLowerCase().replace(/[^a-z0-9]/g, '-');
-      const sanitizedUsername = member.user.username.toLowerCase().replace(/[^a-z0-9]/g, '-');
-      const channelName = `${savedTicket.id}_${sanitizedDisplayName}_${sanitizedUsername}`.substring(0, 100);
-
-      // get the staff/admin roles from the database
-      const rolePerms = await staffRoleRepo
-        .createQueryBuilder()
-        .select(['type', 'role'])
-        .where('guildId = :guildId', { guildId: guildId })
-        .getRawMany();
-
-      // Extract role IDs from mentions
-      const staffRoleIds = rolePerms
-        .map(role => extractIdFromMention(role.role))
-        .filter((id): id is string => {
-          if (!id) {
-            enhancedLogger.warn(`Invalid role format encountered`, LogCategory.COMMAND_EXECUTION, {
-              guildId,
-            });
-            return false;
-          }
-          return true;
-        });
-
-      // Use utility function to create permissions
-      const permOverwrites = createPrivateChannelPermissions(
-        guildId,
-        [member.id],
-        staffRoleIds,
-        PermissionSets.TICKET_CREATOR,
-      );
-
-      // create the channel with all perms
-      const channel = await guild.channels.create({
-        name: channelName,
-        type: 0, // text channel
-        parent: category, // category
-        permissionOverwrites: permOverwrites,
-      });
-
-      await interaction.reply({
-        content: `${lang.ticket.created}${channel}`,
-        flags: [MessageFlags.Ephemeral],
-      });
-
-      // send ticket welcome message with @ mention
-      const welcomeMsg = `<@${member.user.id}>\n\n${lang.ticket.welcomeMsg}`;
-      const buttonOptions = new ActionRowBuilder<ButtonBuilder>().setComponents(
-        new ButtonBuilder().setCustomId('admin_only_ticket').setLabel('Admin Only').setStyle(ButtonStyle.Secondary),
-        new ButtonBuilder().setCustomId('close_ticket').setLabel('Close Ticket').setStyle(ButtonStyle.Danger),
-      );
-      const descriptionMsg = `${description}`;
-      const newChannel = channel as TextChannel;
-
-      const welcome = await newChannel.send({
-        content: welcomeMsg,
-        components: [buttonOptions],
-      });
-      await newChannel.send(`\u200B\n${descriptionMsg}`);
-
-      // Check if staff should be pinged on ticket creation
-      const botConfig = await botConfigRepo.findOneBy({ guildId });
-      if (botConfig?.enableGlobalStaffRole && botConfig?.globalStaffRole) {
-        let shouldPingStaff = false;
-
-        if (isLegacyType) {
-          // Check legacy type ping setting from TicketConfig
-          const pingColumn = resolveLegacyPingColumn(ticketType);
-          if (pingColumn && modalTicketConfig) {
-            shouldPingStaff = modalTicketConfig[pingColumn] as boolean;
-          }
-        } else {
-          // Reuse resolved custom type config fetched earlier
-          shouldPingStaff = resolved.customType?.pingStaffOnCreate ?? true;
-        }
-
-        if (shouldPingStaff) {
-          await newChannel.send({
-            content: `${botConfig.globalStaffRole}\n📨 A new **${displayName}** ticket has been created!`,
-          });
-        }
-      }
-
-      await ticketRepo.update(
-        { id: savedTicket.id, guildId },
-        {
-          messageId: welcome.id,
-          channelId: newChannel.id,
-          status: 'opened',
-        },
-      );
-
-      enhancedLogger.info(`Ticket created: #${savedTicket.id} (${ticketType})`, LogCategory.COMMAND_EXECUTION, {
-        userId: interaction.user.id,
-        guildId,
-        ticketId: savedTicket.id,
-        ticketType,
-        channelId: newChannel.id,
-      });
-    } catch (error) {
-      enhancedLogger.error(
-        'Failed to create ticket',
-        error instanceof Error ? error : new Error(String(error)),
-        LogCategory.COMMAND_EXECUTION,
-        {
-          userId: interaction.user.id,
-          guildId,
-          ticketType,
-        },
-      );
-      await interaction.reply({
-        content: lang.ticket.error,
-        flags: [MessageFlags.Ephemeral],
-      });
-      return;
-    }
-  }
-
-  /* MAKING A TICKET ADMIN ONLY */
-  if (interaction.isButton() && interaction.customId === 'admin_only_ticket') {
-    enhancedLogger.debug(`Button: admin_only_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-
-    // build a confirmation message with buttons
-    const confirmRow = new ActionRowBuilder<ButtonBuilder>().addComponents(
-      new ButtonBuilder().setCustomId('confirm_admin_only_ticket').setLabel('Confirm').setStyle(ButtonStyle.Danger),
-      new ButtonBuilder().setCustomId('cancel_admin_only_ticket').setLabel('Cancel').setStyle(ButtonStyle.Secondary),
-    );
-
-    await interaction.reply({
-      content: lang.ticket.adminOnly.confirm,
-      components: [confirmRow],
-      flags: [MessageFlags.Ephemeral],
-    });
-  }
-  if (interaction.isButton() && interaction.customId === 'confirm_admin_only_ticket') {
-    enhancedLogger.debug(`Button: confirm_admin_only_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-    await interaction.update({
-      content: lang.ticket.adminOnly.changing,
-      components: [],
-    });
-    await ticketAdminOnlyEvent(client, interaction);
-  }
-  if (interaction.isButton() && interaction.customId === 'cancel_admin_only_ticket') {
-    enhancedLogger.debug(`Button: cancel_admin_only_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-    await interaction.update({
-      content: lang.ticket.adminOnly.cancel,
-      components: [],
-    });
-  }
-
-  /* CLOSING A TICKET */
-  if (interaction.isButton() && interaction.customId === 'close_ticket') {
-    enhancedLogger.debug(`Button: close_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-
-    // build a confirmation message with buttons
-    const confirmRow = new ActionRowBuilder<ButtonBuilder>().addComponents(
-      new ButtonBuilder().setCustomId('confirm_close_ticket').setLabel('Confirm Close').setStyle(ButtonStyle.Danger),
-      new ButtonBuilder().setCustomId('cancel_close_ticket').setLabel('Cancel').setStyle(ButtonStyle.Secondary),
-    );
-
-    await interaction.reply({
-      content: lang.ticket.close.confirm,
-      components: [confirmRow],
-      flags: [MessageFlags.Ephemeral],
-    });
-  }
-  if (interaction.isButton() && interaction.customId === 'confirm_close_ticket') {
-    enhancedLogger.debug(`Button: confirm_close_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-    await interaction.update({
-      content: lang.ticket.close.closing,
-      components: [],
-    });
-    await ticketCloseEvent(client, interaction);
-  }
-  if (interaction.isButton() && interaction.customId === 'cancel_close_ticket') {
-    enhancedLogger.debug(`Button: cancel_close_ticket`, LogCategory.COMMAND_EXECUTION, {
-      userId: interaction.user.id,
-      guildId,
-    });
-    await interaction.update({
-      content: lang.ticket.close.cancel,
-      components: [],
-    });
-  }
+  await dispatchTicketInteraction(client, interaction);
 };
